vyn_iq/requirements.txt
asgiref==3.11.0
Django==5.2.9
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
greenlet==3.2.4
playwright==1.55.0
pyee==13.0.0
PyJWT==2.10.1
sqlparse==0.5.4
typing_extensions==4.15.0
